requirements.txt
pytest>=8.3

python-dotenv=1.2.1